Merged various STM32F0 files for a combined project with all necesary libraries.
Using the SPI_DMA project as a base, I added functionality for two
timers, one for a PWM output and one for a sampling clock. ADC was also
added. Tidyied up the main loop with some initialisation functions. Also
changed the clock settings so it should work on the PCB.